シードSQLのgenerate_series呼び出しに型キャストを付け、EXPLAIN取得を可能にする
pg_stat_statementsがリテラル引数も$Nパラメータに正規化するため、
nrpostgresqlreceiver(top_query_collection)がEXPLAIN実行時にgenerate_seriesの
オーバーロード解決に失敗し"function generate_series(unknown, unknown) is not
unique"で失敗していた。

+++ db/init/04-init-assertion-rules.sql
@@ -38,5 +38,5 @@ SELECT
     1,
     true,
     company_id::text
-FROM generate_series(1, 50) AS company_id
+FROM generate_series(1::integer, 50::integer) AS company_id
 ON CONFLICT (id) DO NOTHING;
